fixed: moved tests files as .spec.ts in src
kepts tests for e2e
